frontend+backend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai==1.54.0
httpx==0.27.2
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
mcp==1.2.0
langchain-mcp-adapters==0.1.0
